+++ ios/ViewManagers/SmileIDEnhancedDocumentVerificationViewManager.swift
@@ -32,21 +32,20 @@ class SmileIDEnhancedDocumentVerificationViewManager: SmileIDBaseViewManager {
         if let consentInformation = params["consentInformation"] as? NSDictionary{
           self.product.consentInformation = consentInformation.toConsentInfo()
         } else  {
-          self.product.consentInformation = ConsentInformation(
-            consented: ConsentedInformation(
-              consentGrantedDate: getCurrentIsoTimestamp(),
-              personalDetails: false,
-              contactInformation: false,
-              documentInformation: false
-            )
-          )
+          self.product.consentInformation = nil
+        }
+        
+        var autoCapture: String?
+        if let autoCaptureValue = params["autoCapture"] as? String {
+          autoCapture = autoCaptureValue
         }
 
         self.product.extraPartnerParams = params["extraPartnerParams"] as? [String: String] ?? [:]
         self.product.userId = params["userId"] as? String
         self.product.jobId = params["jobId"] as? String
         self.product.countryCode = countryCode
-        self.product.enableAutoCapture = params["enableAutoCapture"] as? Bool ?? true
+        self.product.autoCaptureTimeout = params["autoCaptureTimeout"] as? Int ?? 10
+        self.product.autoCapture = autoCapture?.toAutoCapture() ?? .autoCapture
         self.product.allowAgentMode = params["allowAgentMode"] as? Bool ?? false
         self.product.allowNewEnroll = params["allowNewEnroll"] as? Bool ?? false
         self.product.showAttribution = params["showAttribution"] as? Bool ?? true